fix(sdlc): extract shared resolveBranch and validate numeric args
Move resolveBranch() to lib/state.js to eliminate code duplication.
Add NaN validation for --total-tasks and --wave arguments in
execute-state.js with specific error messages.
